% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportrxiv_word.R
\name{sportrxiv_word}
\alias{sportrxiv_word}
\title{SportRxiv Word document}
\usage{
sportrxiv_word(fig_caption = TRUE, md_extensions = NULL, ...)
}
\arguments{
\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{md_extensions}{Markdown extensions to be added or removed from the default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for additional details.}

\item{...}{Further arguments to pass to \code{\link[rmarkdown]{word_document}}.}
}
\description{
Template for creating an article according to SportRxiv guidelines in Word..
}
\details{
To be added
}
\section{Functions}{
\itemize{
\item \code{sportrxiv_word}: create .docx-files. \code{class} parameter is ignored. \emph{This function
should be considered experimental.}
}}

\examples{
NULL
}
\seealso{
\code{\link[rmarkdown]{word_document}}
}
